+++ Compilador/lib/parser/asignacion.rs
@@ -8,6 +8,7 @@ use crate::scanners::ws::*;
 use crate::scanners::id::*;
 use crate::parser::reglas_expresion::exp::*;
 use crate::parser::reglas_expresion::valor::*;
+use crate::parser::dimensiones::*;
 use crate::semantica::globales::*;
 use crate::semantica::tabla_variables::*;
 
@@ -59,90 +60,16 @@ pub fn asignacion_interna(input: &str) -> IResult<&str, &str> {
     Err(_) => next
   };
 
-
   let mut hay_dims = false;
   let variable = buscar_variable(id_valor);
-  next = match corchete(next) {
+  next = match con_dim(id_valor)(next) {
     Ok((next_input, _)) => {
-      hay_dims = true;
-      match variable.dimensiones.len() {
-        0 => {
-          println!("Variable no tiene dimensiones");
-        },
-        _ => {
-          PILA_DIMENSIONES.lock().unwrap().push((variable.clone(), 1));
-          PILA_OPERADORS.lock().unwrap().push("(".to_owned());
-        }
-      };
-
-      match tuple((delimited(ws, exp, ws), tag("]")))(next_input) {
-        Ok((next_i, _)) => {
-          {
-            PILA_DIMENSIONES.lock().unwrap().pop();
-            PILA_OPERADORS.lock().unwrap().pop();
-          }
-          let mut pila_valores = PILA_VALORES.lock().unwrap();
-          let valor = pila_valores.pop().unwrap();
-          let mut cuadruplos = CUADRUPLOS.lock().unwrap();
-          let mut constantes = CONSTANTES.lock().unwrap();
-          cuadruplos.agregar_cuadruplo_verificar(valor.direccion, variable.dimensiones[0]);
-          drop(pila_valores);
-          match variable.dimensiones.len() {
-            1 => {
-              let dir = constantes.agregar_constante(variable.direccion.to_string(), variable.tipo.clone());
-              cuadruplos.agregar_cuadruplo("+",valor.clone(), dir.clone());
-              next_i
-            },
-            2 => {
-              let dim_constante = constantes.agregar_constante(variable.dimensiones[1].to_string(), variable.tipo.clone());
-              cuadruplos.agregar_cuadruplo("*", valor.clone(), dim_constante.clone());
-              match corchete(next_i) {
-                Ok((next_input, _)) => {
-                  {
-                    PILA_DIMENSIONES.lock().unwrap().push((variable.clone(), 2));
-                    PILA_OPERADORS.lock().unwrap().push("(".to_owned());
-                  }
-                  drop(cuadruplos);
-                  drop(constantes);
-                  match tuple((delimited(ws, exp, ws), tag("]")))(next_input) {
-                    Ok((next_i, _)) => {
-                      println!("exp 2");
-                      {
-                        PILA_DIMENSIONES.lock().unwrap().pop();
-                        PILA_OPERADORS.lock().unwrap().pop();
-                      }
-                      let mut pila_valores = PILA_VALORES.lock().unwrap();
-                      let mut cuadruplos = CUADRUPLOS.lock().unwrap();
-                      let mut constantes = CONSTANTES.lock().unwrap();
-                      let valor = pila_valores.pop().unwrap();
-                      cuadruplos.agregar_cuadruplo_verificar(valor.direccion, variable.dimensiones[1]);
-                      drop(pila_valores);
-                      let dir = constantes.agregar_constante(variable.direccion.to_string(), variable.tipo.clone());
-                      cuadruplos.agregar_cuadruplo("+", valor.clone(), dir.clone());
-                      next_i
-                    },
-                    Err(err) => {
-                      println!("Error mero dentro");
-                      return Err(err);
-                    }
-                  }
-                },
-                Err(_) => next_i
-              }
-            },
-            _ => next_i
-          }
-        },
-        Err(err) => {
-          println!("Error no sé");
-          return Err(err);
-        }
+      if next_input == next {
+        hay_dims = true;
       }
+      next_input
     },
-    Err(_) => {
-      PILA_VALORES.lock().unwrap().push(variable);
-      next
-    }
+    Err(err) => return Err(err)
   };
 
   next = match tuple((ws, tag("="), ws, exp))(next) {
